move command from listview to list
