+++ ignite.json
@@ -3,6 +3,7 @@
     "component",
     "container",
     "listview",
+    "list",
     "redux",
     "saga",
     "screen"
